dcoops/requirements.txt
aiohttp==3.7.4
alembic==1.4.3
async-timeout==3.0.1
attrs==20.3.0
cffi==1.14.4
chardet==3.0.4
cryptography==3.4.7
discord.py[voice]==1.7.3
idna==2.10
Mako==1.1.3
MarkupSafe==1.1.1
multidict==4.7.6
pycparser==2.20
PyMySQL==0.10.1
PyNaCl==1.3.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.4.26
yarl==1.5.1
youtube-dl==2020.12.22
wheel==0.36.2
beautifulsoup4==4.9.3
discord-pretty-help==1.3.2
gtts==2.2.3
pika==1.2.0